pyproject.toml
[tool.poetry]
name = "24.02.23newsparser-telegrambot"
version = "0.1.0"
description = ""
authors = ["Ariulis <andrey59995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.25.1"
requests-html = "^0.10.0"
python-dotenv = "^1.0.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
